libdrgn: dwarf_info: skip DW_CFA_GNU_args_size
Stephen reported "unknown DWARF CFI opcode 0x2e" errors. 0x2e is
DW_CFA_GNU_args_size, which "takes an unsigned LEB128 operand
representing an argument size" that "specifies the total of the size of
the arguments which have been pushed onto the stack". It seems fairly
common in .eh_frame. We have no use for this information, so we can
simply skip it (elfutils and LLDB apparently do the same).

Closes #480.

Reported-and-tested-by: Stephen Brennan <stephen.s.brennan@oracle.com>
Signed-off-by: Omar Sandoval <osandov@osandov.com>