cli: fix 'staticmethod' object is not callable for Python < 3.10
Prior to Python 3.10, the object returned by @staticmethod apparently
cannot be called unless bound to the class instance. So the new changes
in the fixed commit cause the CLI to fail on Python < 3.10.

We can resolve this in a variety of ways, but the one which doesn't
result in any static checker warnings is the simplest: just move this
function out to a module-level function. There are no users in
sub-classes which need to be converted.

Fixes: 4a6a9f33 ("cli: allow enabling/disabling debug info finders with --{try,no}-symbols-by")

Signed-off-by: Stephen Brennan <stephen.s.brennan@oracle.com>