cli: handle !isatty(stdin)
Redirecting drgn's stdin (e.g., to a pipe or a file) is really broken:
with the enhanced REPL, it fails with "termios.error: (25,
'Inappropriate ioctl for device')" and "AttributeError: 'UnixConsole'
object has no attribute '_UnixConsole__buffer'". With the basic REPL, it
prints ps1/ps2 prompts and expects extra newlines to terminate indented
blocks.

The standard Python REPL handles this by checking if stdin is a
terminal. Unfortunately, code.interact()/_pyrepl don't do this, so let's
do the check and exec manually.

Reported-by: Stephen Brennan <stephen.s.brennan@oracle.com>
Reported-by: Jeff Layton <jlayton@kernel.org>
Signed-off-by: Omar Sandoval <osandov@osandov.com>