libdrgn: linux_kernel: disable debuginfod except on Fedora
My debuginfod optimizations for the Linux kernel haven't been deployed
anywhere other than Fedora's debuginfod servers. On other servers, it's
better to not even try since downloads will take forever. Do a sketchy
check to determine whether the kernel is from Fedora and disable the
debuginfod finder if not. It can still be reenabled manually with
--try-symbols-by debuginfod or by a plugin.

Signed-off-by: Omar Sandoval <osandov@osandov.com>