CI: install pyroute2 < 0.9.1 for Python < 3.9.
pyroute 0.9.1 dropped support for Python < 3.9. On Python 3.6 and 3.7,
it fails with an ImportError from typing that the try/except in
tests/linux_kernel/helpers/test_tc.py catches by accident, but on 3.8 it
fails with a TypeError: 'type' object is not subscriptable. Fix it by
installing an older version of pyroute2 on those Python versions.

Signed-off-by: Omar Sandoval <osandov@osandov.com>