Add absence reason to absent objects
An object can be absent for multiple reasons: e.g., the object is a
placeholder, it was optimized out, or its value is represented using a
DWARF opcode we don't support. We print all of these as "<absent>",
which is technically correct but criminally vague.

Add an absence reason to struct drgn_object/drgn.Object. For now, this
only has one possible value, "OTHER", but it will be made more specific
in follow-up commits.

Signed-off-by: Omar Sandoval <osandov@osandov.com>