libdrgn: debug_info: index modules by name
We currently index modules by their unique key (kind, name, info). I've
wanted to look up modules using only their name, which can't be done
efficiently with the current module table. Instead, let's make the table
map a name to a list of modules with that name. It's very rare for
modules to have the same name, so this is just as efficient for the
common case of existing operations and enables us to add a name lookup.

Signed-off-by: Omar Sandoval <osandov@osandov.com>