Add module lookups by name
It's rare to know the address of the module you're looking for, so the
existing module lookup functions are often not helpful. Add
drgn_module_find_by_name() and drgn.Program.module(name) to address
this. Userspace still has the issue that you may not know the absolute
path of the binary you care about, but this at least helps the kernel.

Signed-off-by: Omar Sandoval <osandov@osandov.com>