tests: don't shell out for CLI tests
In some test environments, like Buck, it's not easy to reexecute the
Python interpreter and have it find the drgn module. Instead of using
subprocess, fork and call drgn.cli._main() directly (setting up the
necessary pipes ourselves). This gives us slightly less realistic
end-to-end coverage, but it makes the tests more robust.

Signed-off-by: Omar Sandoval <osandov@osandov.com>