libdrgn: x86_64: fix restoring rbp for ORC_TYPE_REGS
Our handling of rbp in ORC_TYPE_REGS is incorrect. First, we set the CFI
rule to restore rbp from the on-stack pt_regs. Then, we set it again
based on bp_reg, which is always ORC_REG_UNDEFINED, meaning to use the
previous rbp. The intention is to keep the value we just got from
pt_regs, but we interpret it as the previous frame's rbp, which is
incorrect.

Fix it by skipping the bp_reg handling for ORC_TYPE_REGS.

Reported-by: Leo Martins <leomar@meta.com>
Signed-off-by: Omar Sandoval <osandov@osandov.com>