tests: fix deserialize_struct64 tests on i386
struct foo64 is only 20 bytes with i386's alignment requirements, so we
get the following compiler warning on i386:

  ‘memcpy’ forming offset [20, 23] is out of the bounds [0, 20] of object ‘foo’ with type ‘struct foo64’

and a crash when running the tests. Fix it by adding padding manually.

I verified that this only affects the test cases. All of the structs we
actually use deserialize_struct64 for are properly defined.

Fixes #493.

Signed-off-by: Omar Sandoval <osandov@osandov.com>