libdrgn: python: add missing garbage collection support
As Stephen Brennan discovered in #498, we're missing
Py_TPFLAGS_HAVE_GC/tp_traverse in a bunch of places. I misunderstood
tp_clear not being needed for immutable types to mean that tp_traverse
isn't needed, either, but that's wrong. We're also missing
PyObject_GC_UnTrack() calls in the tp_dealloc implementations for
garbage collected types. Add the missing implementations/calls.

I tested this with a script that runs tests and compares the set of live
programs (which we save for our logging hack):
https://gist.github.com/osandov/b4cdab8e61476dd5cfa306dab5d4c7cc. A few
test classes needed a tearDownClass() method to clear the program, but
after doing that, there are no leaks. (There appears to be a false
positive for our expectedFailure() case, but that should hopefully be
gone soon.)

Benchmarks also showed no significant change in performance from the
added calls.

Signed-off-by: Omar Sandoval <osandov@osandov.com>