libdrgn: elf_symtab: fix overriding partial table with full table
If we find a partial symbol table (i.e., .dynsym) and then later add a
file with a full symbol table (i.e., .symtab), then we're supposed to
override the partial table with the full table. However, we have the
condition in find_module_elf_symtab() flipped. Fix it and add a test.

Signed-off-by: Omar Sandoval <osandov@osandov.com>