drgn.helpers.linux: add timekeeping helpers
The monotonic, real, and boot timestamps of a crash are all useful
information. Add helpers for accessing them (and International Atomic
Time for completeness), both in seconds and nanoseconds.

Exact timestamps can only be read using the hardware clocksource, so we
can only get coarse timestamps, which are updated in memory on each
tick.

ktime_t is a nuisance thanks to it being a union before Linux 4.10, so
we don't provide any ktime_t helpers.

Also add uptime() and uptime_pretty() helpers as convenient shortcuts.

Signed-off-by: Omar Sandoval <osandov@osandov.com>