pre-commit: update flake8, mypy, and pre-commit-hooks
These are the last versions of flake8, and mypy that support Python 3.8.
We're already on the last versions of black and isort that support
Python 3.8.

The new mypy apparently doesn't understand a couple of things in
takes_program_or_default(), so add a couple of ignores. Also remove a
couple that are no longer necessary.

Signed-off-by: Omar Sandoval <osandov@osandov.com>