Separate debug info options for directories and debug link directories
Stephen reported that --debug-directory does not appear to accept
relative paths. This is because relative paths are only used for
searches by debug link, where they are interpreted as relative to the
directory containing the loaded file. I copied this behavior from
libdwfl, but it is very unlikely to be what the user wants. In my
opinion, it's surprising enough to warrant changing the semantics of the
directories option.

This splits the directories used for searches by debug link into a
separate option, debug_link_directories. Instead of implicitly appending
or prepending the directory containing the loaded file depending on
whether the directory option is absolute or relative, make it explicit:
the string "$ORIGIN" (or "${ORIGIN}") is replaced with the absolute path
of the directory containing the loaded file (this is inspired by the
rpath option used by dynamic linkers/loaders). We also support checking
under each directory in the main directories option by specifying an
empty string.

With debug link directories separated out, remove any special treatment
for relative paths in the directories option. The empty string is
forbidden for now just in case we want to give it a special meaning in
the future.

The new default settings are directories=("/usr/lib/debug",),
debug_link_directories=("$ORIGIN", "$ORIGIN/.debug", ""), which is
equivalent to the old default directories=("", ".debug",
"/usr/lib/debug"). Other than the better behavior of --debug-directory,
this should only affect users that were already doing something very
advanced.

Closes #500.

Reported-by: Stephen Brennan <stephen.s.brennan@oracle.com>
Signed-off-by: Omar Sandoval <osandov@osandov.com>