tests: skip timekeeping tests using time.clock_gettime_ns() on Python 3.6
I was originally planning on dropping Python 3.6 and 3.7 support in the
next release, but now I'm planning on doing a small release to get some
fixes out. The only place we broke that support was in a couple of tests
for the new timekeeping helpers that use time.clock_gettime_ns(), which
was added in Python 3.7. We can easily skip those for now to preserve
support for one more release.

Signed-off-by: Omar Sandoval <osandov@osandov.com>