vmtest.vm: use "b" magic sysrq instead of "o"
I've been seeing panics on shutdown in vmtest caused by the sleep
command after the echo o > /proc/sysrq-trigger failing, either with
"command not found" or "Input/output error". Presumably we're racing
with the root filesystem being torn down. This is cosmetic, but it adds
noise to logs. Switch to echo b > /proc/sysrq-trigger, which is
synchronous and therefore doesn't need the subsequent sleep.

Signed-off-by: Omar Sandoval <osandov@osandov.com>