vmtest.vm: replace signal handler with setpriv --pdeathsig
run_in_vm() currently installs a signal handler for SIGTERM just to
forward the signal to the QEMU process. This has a few issues:

1. If the process is killed by another signal (e.g., SIGKILL), that
   doesn't get forwarded to QEMU.
2. signal.signal() only works in the main thread, which would require a
   workaround for running tests in parallel.
3. If there was already a signal handler installed, it is not called or
   restored afterwards.

Setting a parent-death signal with prctl(PR_SET_PDEATHSIG) is an
alternative, but it's difficult to use directly from Python for a couple
of reasons:

1. The standard library doesn't expose prctl(), so it would require
   using ctypes.
2. It needs to be called in subprocess.Popen()'s preexec_fn, which is
   not safe in the presence of threads.

Luckily, util-linux provides setpriv --pdeathsig for this purpose, which
we can use to wrap our QEMU call. It has been supported since util-linux
v2.33 (released in 2018). BusyBox doesn't support it, but all of the
systems we expect to run vmtest on will have a non-ancient version of
util-linux. The usual race condition of the parent dying between fork()
and prctl(PR_SET_PDEATHSIG) thankfully doesn't affect us since in that
case, QEMU will fail to connect to the serial socket and exit anyways.

With that, we can get rid of the signal handler.

Signed-off-by: Omar Sandoval <osandov@osandov.com>