_drgn_util: add argparse help formatter that supports multiple lines
The default argparse.HelpFormatter wraps an entire string as one
paragraph. argparse.RawDescriptionHelpFormatter and
argparse.RawTextHelpFormatter use the given string verbatim. For
upcoming commands, we want a middle ground: multiple paragraphs that can
each be wrapped, as well as blocks of text that are not wrapped (for
tables and example output, for example).

Add a HelpFormatter subclass that does this with a
reStructuredText-inspired syntax:

    This is a paragraph, which will
    be wrapped.

    |This is a line block.
    |It will not be wrapped.

P.S. Lib/argparse.py in CPython currently says about HelpFormatter:

    Only the name of this class is considered a public API. All the methods
    provided by the class are considered an implementation detail.

But we don't have much of a choice, so we do the best we can with unit
tests.

Signed-off-by: Omar Sandoval <osandov@osandov.com>