libdrgn: always create all modules in load_debug_info()
For the Linux kernel at least, it's important to have modules populated
even if there is no debuginfo available, because the modules are
necessary to load the built-in ORC data for unwinding.

The break statement here intentionally prevents exhausting the module
iterator, so that the Linux kernel modules don't get created. It doesn't
impact functionality afterwards: only modules that we want to find debug
info for are added to the module vector. So as far as I can see, this is
only a performance optimization.

The documentation for load_debug_info() only states "Try to load all
debugging information for all loaded modules", which doesn't really
specify whether the drgn Modules are actually created. By dropping this
break statement, we may slightly increase runtime for the --main-symbols
case, but we also improve correctness for those same cases, by ensuring
that stack tracing (via ORC) can work without DWARF debug info.

Closes #505

Signed-off-by: Stephen Brennan <stephen@brennan.io>