Add crash compatibility mode
In order to make it easier for users to migrate from crash to drgn, we
are going to port crash commands to drgn. The killer feature will be a
--drgn option that helps users onboard to drgn's advanced features.

Add a command namespace for crash commands using the new command system
and a few convenience functions for defining crash commands.

See #507.

Signed-off-by: Omar Sandoval <osandov@osandov.com>