tests: fix "can't pickle local object" errors on Python 3.14
Python 3.14 changed the default multiprocessing start method to
"forkserver", which requires that all arguments to Process are
picklable. See python/cpython#125714.

We have a couple of test cases that pass local functions as the process
target, and apparently local functions aren't picklable. This results in
test failures like:

  _pickle.PicklingError: Can't pickle local object <function TestThreads.test_threads.<locals>.proc_func at 0x7f87b0811380>

In both cases, the local function are redundant wrappers around a method
call, so we can fix this by passing the method directly.

Signed-off-by: Omar Sandoval <osandov@osandov.com>