docs: remove validators from user guide
Validators are listed in the "core concepts" section of the user guide,
but they never really caught on and aren't necessary to introduce to new
users. Move the documentation to ValidationError and the examples to
validate_list() and validate_list_for_each_entry().

Taking a step back, I think the current interface for validators is
wrong. It's fine if you know what data structure you want to validate,
but there's no way to do validation for higher-level helpers that use
list_for_each_entry(), etc. internally. A better interface would allow
enabling/disabling validation for *everything*, either with function
calls, context managers, or both. The tricky part is supporting that
without too much overhead.

Regardless, this is the interface we have for now, so let's keep it but
deemphasize it.

Signed-off-by: Omar Sandoval <osandov@osandov.com>