drgn.helpers.linux.kernfs: add kernfs_children() helper
Add a helper, kernfs_children(), that returns an iterator over the
children of a directory in kernfs.

The implementation of this must check kn.flags to ensure it is a
KERNFS_DIR. struct kernfs_node unions dir and attr and without this
check, calling kernfs_children() on a KERNFS_FILE kernfs node may return
corrupt data or trigger a page fault that is confusing to the user.

Signed-off-by: Joanne Koong <joannelkoong@gmail.com>