drgn.helpers.linux.sched: add get_task_state()
This is like task_state_to_char(), but with the additional full name
used in /proc/pid/status. It will be used for porting crash commands
that display a task state (in lieu of the bitmask that crash prints,
which I'd rather not do since we'd have to hardcode many of the bit
macros).

Signed-off-by: Omar Sandoval <osandov@osandov.com>