crash: pipe to pager by default
The default behavior of the crash utility is to use a pager, so let's do
the same. We default to less -E -X, fall back to more if less is not
available, and fall back to no pager if neither is available.

This also needs a small update to _shell_command() to make sure that
pipelines also get sent through the pager.

Signed-off-by: Omar Sandoval <osandov@osandov.com>