drgn.commands: format manual usage strings in rST
For Sphinx documentation of commands, we want extra markup to
bold/italicize options. For in-program help, we don't want the markup.
For argparse-generated usage strings, drgndoc already handles adding the
markup, so this works in that case but not for manual usage strings.

To fix this, make manual usage strings include the rST markup for the
sake of drgndoc, but strip the markup at runtime.

Signed-off-by: Omar Sandoval <osandov@osandov.com>