crash: port set command
This is really two commands in one: setting/showing the current task
context, and setting/showing configuration options. This unfortunate
multiplexing requires us to do some argparse gymnastics. Many of crash's
options are not applicable to drgn. For now, we only support "scroll".

Now that we have a couple of commands that deal with the current context
(set and mount), this also adds a helper function for dealing with the
context in --drgn output (drgn.commands.crash.add_crash_context()). The
mount command will be converted to use it in the next commit.

Signed-off-by: Omar Sandoval <osandov@osandov.com>