libdrgn: generalize drgn_language::bit_offset() to type_subobject()
For the next commit adding a drgn_object_subprogram() operation, we want
to reuse the designator parsing logic that we use for offsetof() to also
get the type and bit field size. drgn_language::bit_offset() already
gets this information but doesn't return it. Replace it with a more
general type_subobject() hook that returns everything and also allows
starting with an array subscript.

Signed-off-by: Omar Sandoval <osandov@osandov.com>