python: raise drgn.ObjectNotFoundError instead of KeyError/LookupError
Add a new drgn.ObjectNotFoundError exception type to use for failed
prog[], prog.object(), prog.variable(), prog.constant(), and
prog.function() lookups. This has a few benefits:

1. It allows catching object lookup errors without the risk of catching
   a stray error from a dict or attribute lookup. See #28.
2. It allows us to have a more descriptive error message. See #527.
3. It makes it possible to get the name that could not be found from the
   exception.

It is a subclass of KeyError (and thus also of LookupError), so this
should be backwards compatible. The implementation is inspired by
AttributeError in CPython.

If we end up liking this, we might want to do the same for object member
lookups.

Signed-off-by: Omar Sandoval <osandov@osandov.com>