crash: port p command (without expression evaluation)
The most basic use case of the p command is printing a single object,
which we can easily support. We can also support members and array
subscripts fairly easily. Arbitrary expressions will be a lot more work,
so we'll come back to that later.

This command is in the _eval module with the expectation that it'll
share expression evaluation code with the eval command when that is
added. The underscore is to prevent name conflicts like the one fixed in
6067b21ee58d ("Rename drgn/commands/_builtin/crash/sys{,tem}.py").

Signed-off-by: Omar Sandoval <osandov@osandov.com>