drgn.commands: use argparse to generate help for custom commands
The crash eval command has slightly non-standard syntax, so it needs to
be a custom command. However, it's close enough to a normal command that
we'd like to have help formatted like argparse. In fact, most custom
commands can have their help expressed with argparse, so let's always
use argparse to format help even when we don't use it to parse anything.

Signed-off-by: Omar Sandoval <osandov@osandov.com>