tools.fsrefs: handle Btrfs block device holder change in Linux 6.17
Linux kernel commit 40426dd147ff ("btrfs: use the super_block as holder
when mounting file systems") (in v6.17) made it so that btrfs_fs_type is
no longer the bdev holder. Instead of checking for this, use a more
general method of iterating over all super blocks and checking for
Btrfs.

Signed-off-by: Omar Sandoval <osandov@osandov.com>