commands: builtin: add py command
In #537 it was pointed out that the ability to pipe output produced by
executing Python statements would be very useful. Unfortunately the
shell redirection operators are part of the Python grammar as well, and
there are many cases of ambiguity, where a command could be split into
python code and shell pipeline in multiple valid ways.

However, these ambiguities may not be a dealbreaker. We can resolve them
by always splitting on the first shell operator which produces a valid
Python code on the left hand side. In cases where you want to force a
different interpretation, you can wrap your Python code in parentheses.
These ensure that any shell operator within the parentheses doesn't
introduce a pipeline, because the code prior to them is incomplete
without the closing parenthesis.

Signed-off-by: Stephen Brennan <stephen.s.brennan@oracle.com>