vmtest.config: disable CONFIG_SLUB_CPU_PARTIAL for alternative flavor
Ever since SLAB was removed, the alternative flavor isn't providing any
extra coverage for slab allocator options. The most relevant option for
upcoming slab helpers that isn't covered is CONFIG_SLUB_CPU_PARTIAL,
which is on by default. Turn it off for the alternative flavor so we can
test it.

Signed-off-by: Omar Sandoval <osandov@osandov.com>