tests: fix fork_and_stop() returning while process is still on CPU
There is a window between a process being flagged as stopped and it
actually descheduling. Various stack tracing tests have been flaky with
"cannot unwind stack of running task" errors due to catching the process
in this window.

Fix it by waiting for /proc/pid/syscall to not return "running" (which
is what we did before the fixes commit, but now we don't need to check
for a specific syscall number).

Fixes: bab4f43d68aa ("tests: replace fork_and_sigwait() and fork_and_call() with fork_and_stop()")
Signed-off-by: Omar Sandoval <osandov@osandov.com>