vmtest: add vmtest.chroot to easily enter chroot
It can be nice to modify or play around in the chroots after creation.
For example, to install new packages without rebuilding. While there is
a tool for running commands in a vmtest VM, the VMs have no network
access so they're less flexible. While the necessary command isn't
really all that complicated, it's nice to not have to think of it. Add a
new script to enter the rootfs.

Signed-off-by: Stephen Brennan <stephen.s.brennan@oracle.com>