drgn.helpers.linux.kallsyms: make module_kallsyms() public
I recently used _module_kallsyms() to selectively use kallsyms for
certain modules. I promised in #388 that if we found use cases for any
of the internal functions, we could make them public, so let's do that
for module_kallsyms().

This also reorders the functions so they appear in a more natural order
in the Sphinx documentation, but there are no code changes other than
the rename.

Signed-off-by: Omar Sandoval <osandov@osandov.com>