tests: linux_kernel: panic within NMI on x86_64
The test kernel module currently panics from an IRQ handler to test that
drgn can unwind through it. While the unwinder code doesn't necessarily
need to be tested on NMI in addition, it will be helpful for the crash
"bt" command to be tested on more complex stacks. So, trigger an NMI
within the IRQ that will do the actual panic, giving us another layer of
stacks.

Signed-off-by: Stephen Brennan <stephen.s.brennan@oracle.com>