drgn.helpers.linux: add kernel-specific stack helpers
Kernel stacks may span many contexts, each interrupting each other. Drgn
can unwind through most of these transitions, but even so, it doesn't
provide information to the user about them. The crash "bt" command will
need this information in order to highlight the boundary of each stack
transition.

So, add helpers that can inform the user about the execution context for
each part of the stack.  These helpers involve kernel & architecture
specific heuristics, which we keep hidden for easy updates. The support
is best effort: there are some transitions for which drgn doesn't detect
that a frame is "interrupted", and there are stacks which we cannot
categorize. But this is good enough to provide a basic level of support
for the crash "bt" command.

Signed-off-by: Stephen Brennan <stephen.s.brennan@oracle.com>