commands: crash: implement bt command
The command prints stack segments, registers, and it can be called with
a CPU, task, PID, or for all CPUs. Not all crash options are supported:
most notably missing are the options which print (and optionally
annotate) the stack memory. These should be implemented together with
the "rd" command due to their similarity.

We do add two drgn-specific options. First is a "-d" option to format
stack frames as drgn would, rather than as crash does. This is a nice
way for drgn users to feel more comfortable, while still gaining the
benefits of the header, stack segmentation, register dumps, etc.

Second is a "-V" option to print local variable values. This is an
option that crash users would certainly want to have, if it were
available! Including it in this command could be a great carrot for
advertising drgn to existing crash users.

Beyond these two options, there are a few known differences in the
output:

1. The register dumps are missing registers which are part of the
"pt_regs" object, but not part of the architecture registers -- a good
example would be "ORIG_RAX" on x86_64.

2. The stack memory adresses (shown in brackets []) are offset by one
word from what crash shows, because drgn reports the stack pointer prior
to the function return address was pushed to the stack.

3. Drgn reports inline functions. While we could strictly emulate crash
and omit them, the information is useful enough that it is worth
breaking the compatibility. In place of the stack address, we report the
text "(inline)" so these are easy to see.

4. Filenames reported by StackFrame.source() are not absolute paths,
while those reported by crash are. Thus, the filenames included in "bt
-l" are much shorter in drgn.

Closes #538.

Signed-off-by: Stephen Brennan <stephen.s.brennan@oracle.com>