vmtest.enter_kdump: execute command in kdump
The behavior of vmtest is such that whichever command was specified to
launch the VM, will be the same command that is run when rebooting into
the kdump environment.

Sometimes for testing it's most efficient to go directly into the kdump
environment. You might try doing this:

    python -m vmtest.vm --load-test-kmod -k KERNEL python3 -m vmtest.enter_kdump

But it won't work, because once in the kdump environment, the
vmtest.enter_kdump command is executed again.

We can work around this by assuming that the argv are a command to
execute if we're already in the kdump environment. The result is that
you can do:

    python3 -m vmtest.vm --load-test-kmod -k KERNEL python3 -m vmtest.enter_kdump bash

And directly get a bash shell within the kdump environment. Or better
yet:

    python3 -m vmtest.vm --load-test-kmod -k KERNEL python3 -m vmtest.enter_kdump sh -c 'python3 -m drgn -c /proc/vmcore -s $DRGN_TEST_KMOD'

To go directly to a drgn shell within the kdump environment.

Signed-off-by: Stephen Brennan <stephen.s.brennan@oracle.com>