libdrgn: stack_trace: use fallback when ret_addr_regno can't be unwound
During the stack unwind for aarch64, near the bottom of IRQ stack
frames, drgn reaches a point where it can identify a CFI rule which has
rows for many registers, but it can only evaluate one: the rule for
ra_sign_state. If no rule had been found, drgn would fall back to the
architecture's fallback unwinder. If a rule was found, but no registers
could be determined, then drgn would stop the unwind. In this case,
since one register could be determined, drgn continues with the
resulting register state.

It makes sense to not continue unless drgn can unwind ret_addr_regno at
a minimum. The fallback unwinder may be able to get further. However,
there's a catch. According to Omar, a rule omitting a row for the return
address register signifies the end of the stack trace. So, we catch the
case where a rule was included, but no value could be determined, and
return &drgn_not_found, which signals to try the fallback unwinder.

Signed-off-by: Stephen Brennan <stephen.s.brennan@oracle.com>