vmtest: aarch64: add patches to reserve [_text, _stext)
On aarch64, unwinding IRQ stacks in the vmtest kernels can sometimes
fail, because the IRQ stacks may be mapped in the range [_text, _stext).
This is problematic both because (a) /proc/vmcore cannot read those
addresses correctly, and (b) drgn symbolizes addresses in that range as
within the "_text" symbol. Omar fixed this upstream with the patch
"arm64: map [_text, _stext) virtual address range
non-executable+read-only" (currently queued in linux-next for v6.18).

Apply this patch and backport it to all vmtest kernels since v5.11,
which was the first version to have this behavior. The backport has only
one significant change, in v6.8, to deal with a major rewrite of arm64
initial page table setup. It is necessary to create a separate vma for
[_text, _stext) in order to avoid contiguous mappings, so that the later
remap can safely change permissions.

[1]: https://lore.kernel.org/all/20240214122845.2033971-45-ardb+git@google.com/

Signed-off-by: Stephen Brennan <stephen.s.brennan@oracle.com>