tests: linux_kernel: fix test_pgtable_l5_enabled
Recently, kernel 4.14's vmtests have intermittently started failing on
test_pgtable_l5_enabled with "AssertionError: False != True". This
started after f365f33c ("CI: use KVM") - which means that we're now
using the host CPU rather than an emulated one.

It appears that some Github CI systems have the LA57 (5-level paging)
feature enabled. Further, it seems that in 4.14, the LA57 flag is not
cleared in /proc/cpuinfo when 5-level paging is not in use. So we need a
better way to detect 5-level paging being enabled.

The userspace address range is expanded with 5-level paging. So we can
attempt to mmap an address which requires more than 48 VA bits. As long
as we use MAP_FIXED, or we verify that the hinted address range is not
used, we should get the hinted address if supported. If we get the high
address, then we must have 5-level paging enabled. Otherwise, it must be
disabled or unsupported.

Signed-off-by: Stephen Brennan <stephen.s.brennan@oracle.com>