Add drgn.Type.unaliased()/drgn.Type.unaliased_kind()/drgn_qualified_type_unaliased()
Internally in libdrgn, we have drgn_underlying_type(), which we also
open-code in a couple of places in Python. That function drops
qualifiers, so exposing it as-is is kind of error-prone, and the word
"underlying" is a bit overloaded. Instead, let's add
drgn_qualified_type_unaliased() that preserves qualifiers, and expose it
as drgn.Type.unaliased(). Also, some of the existing use-cases only care
about the underlying kind, so also add drgn.Type.unaliased_kind() as a
shortcut/optimization. Then, use both where appropriate.

(I also considered the name "canonical", but that names connotes that
all typedefs would be removed, not just top-level ones.)

Signed-off-by: Omar Sandoval <osandov@osandov.com>