drgn.helpers.linux.cgroup: clean up css_next_child()
Now that we've diverged from the structure of the kernel code for the
CSS_RELEASED case, it doesn't make sense to do the container_of() in the
other cases only to undo it for the head check. Also replace the
pos.serial_nr.value_() with .read_() since serial_nr will get converted
back to an Object for the comparison anyways, and rename next_ to css in
the CSS_RELEASED case now that it is distinct.

Signed-off-by: Omar Sandoval <osandov@osandov.com>