drgn.helpers.linux.mm: add for_each_valid_page_range() helper
Now that we have the necessary SPARSEMEM and FLATMEM infrastructure, we
can add a helper for getting ranges of valid struct pages. This has a
few use cases:

* Fixing #27, a long-standing issue with for_each_page() returning
  invalid pages.
* The crash kmem -p command.
* Adding page identification to identify_address().
* Big batched reads of pages.

Returning the PFN ranges instead of individual PFNs+pages is less
convenient for simpler use cases but especially useful for the latter
two.

Signed-off-by: Omar Sandoval <osandov@osandov.com>