tests: fix test_d_path_dentry_only() on Linux 6.18
Since Linux kernel commit 09a1b33c080f ("preparations to taking
MNT_WRITE_HOLD out of ->mnt_flags") (in v6.18), mounts of a superblock
are added to the head of the list instead of the tail. This means that
d_path(dentry) now uses the latest mount instead of the first mount.
This causes tests.linux_kernel.helpers.test_fs.TestFs.test_d_path_dentry_only
to fail when running tests on a foreign architecture, where we mount the
host rootfs as /host and bind mount the kernel from it at
/lib/modules/$RELEASE. d_path(dentry) has always been documented to use
an arbitrary mount, so it's not worth changing the helper, but let's
change the test to use a directory that hopefully won't be bind mounted.

Signed-off-by: Omar Sandoval <osandov@osandov.com>