drgn.helpers.linux.common: recognize vmapped symbols in identify_address_all()
Modules (and vmlinux on AArch64) are allocated from the vmap region. So,
we should identify them as such (secondarily to their identification as
a symbol). Crash's kmem command recognizes this, too.

Refactor identify_address_all() to always call
_identify_kernel_address(), passing in whether a symbol was found so it
can do the extra vmap lookup. This is done after yielding the
IdentifiedSymbol, so the extra lookup will not be done for
identify_address().

Signed-off-by: Omar Sandoval <osandov@osandov.com>